add login

+++ webcrm/src/main/java/com/rishiqing/dingtalk/webcrm/controller/app/PhoneCallController.java
@@ -3,6 +3,7 @@ package com.rishiqing.dingtalk.webcrm.controller.app;
 import com.rishiqing.dingtalk.biz.http.HttpResult;
 import com.rishiqing.dingtalk.biz.http.HttpResultCode;
 import com.rishiqing.dingtalk.biz.util.LogFormatter;
+import com.rishiqing.dingtalk.isv.api.model.corp.CorpStaffVO;
 import com.rishiqing.dingtalk.isv.api.service.biz.PhoneCallBizService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -40,8 +42,8 @@ public class PhoneCallController {
     @RequestMapping(value = "/callActivateAdmin", method = RequestMethod.POST)
     @ResponseBody
     public Map<String, Object> callActivateAdmin(
-            @RequestParam("corpId") String corpId,
-            @RequestParam(value = "callUserId", required = false) String callUserId
+            HttpServletRequest request,
+            @RequestParam("corpId") String corpId
     ) {
         bizLogger.info(LogFormatter.format(
                 LogFormatter.LogEvent.START,
@@ -49,9 +51,9 @@ public class PhoneCallController {
                 new LogFormatter.KeyValue("corpId", corpId)
         ));
         try{
-            //TODO  这里需要根据权限获取用户信息.这里需要读取当前登录用户的信息，作为主叫方
-            // String loginUserId = callUserId;
-            String result = phoneCallBizService.callActivateAdmin(corpId, callUserId);
+            //这里需要根据权限获取用户信息.这里需要读取当前登录用户的信息，作为主叫方
+            CorpStaffVO loginUser = (CorpStaffVO) request.getAttribute("loginUser");
+            String result = phoneCallBizService.callActivateAdmin(corpId, loginUser.getUserId());
             Map<String, Object> map = new HashMap<>();
             map.put("result", result);
 
@@ -69,16 +71,19 @@ public class PhoneCallController {
     @RequestMapping(value = "/calleeList", method = RequestMethod.POST)
     @ResponseBody
     public Map<String, Object> setCalleeList(
+            HttpServletRequest request,
             @RequestParam("calleeList") String calleeList
     ) {
-        //TODO 这里需要做权限认证
         bizLogger.info(LogFormatter.format(
                 LogFormatter.LogEvent.START,
                 "/setCalleeList",
                 new LogFormatter.KeyValue("calleeList", calleeList)
         ));
         try{
-            //TODO  这里需要根据权限获取用户信息
+            CorpStaffVO loginUser = (CorpStaffVO) request.getAttribute("loginUser");
+            if (loginUser == null) {
+                return HttpResult.getFailure(HttpResultCode.SYS_ERROR.getErrCode(),HttpResultCode.SYS_ERROR.getErrMsg());
+            }
             phoneCallBizService.setCalleeList(calleeList);
             Map<String, Object> map = new HashMap<>();
             map.put("errcode", 0);
@@ -97,6 +102,7 @@ public class PhoneCallController {
     @RequestMapping(value = "/calleeList", method = RequestMethod.DELETE)
     @ResponseBody
     public Map<String, Object> removeCalleeList(
+            HttpServletRequest request,
             @RequestParam("calleeList") String calleeList
     ) {
         bizLogger.info(LogFormatter.format(
@@ -105,7 +111,10 @@ public class PhoneCallController {
                 new LogFormatter.KeyValue("calleeList", calleeList)
         ));
         try{
-            //TODO  这里需要根据权限获取用户信息
+            CorpStaffVO loginUser = (CorpStaffVO) request.getAttribute("loginUser");
+            if (loginUser == null) {
+                return HttpResult.getFailure(HttpResultCode.SYS_ERROR.getErrCode(),HttpResultCode.SYS_ERROR.getErrMsg());
+            }
             phoneCallBizService.removeCalleeList(calleeList);
             Map<String, Object> map = new HashMap<>();
             map.put("errcode", 0);
